+++ firstrun.sh
@@ -16,11 +16,8 @@ echo "Installing required packages."
 read -p "Press any key to continue or ctl-c to abort."
 sudo yum -y install gcc-c++ make gdb automake git lua lua-devel mysql-server mysql-devel java-1.8.0-openjdk boost-devel wget
 sudo chkconfig mysqld on
-#### Test this ######################################
-#
-#sudo mysqld_secure_installation
-#
-####################################################
+sudo service mysqld start
+sudo mysql_secure_installation
 
 #### Install Berkeley DB ####
 echo 'Now installing Berkeley DB 5'
@@ -45,13 +42,17 @@ else
 	mkdir ~/git
 	cd ~/git
 fi
-git clone http://gerrit.swgemu.com/Core3
-git clone http://gerrit.swgemu.com/PublicEngine
+git clone -b unstable http://gerrit.swgemu.com/Core3
+git clone -b master http://gerrit.swgemu.com/PublicEngine
 
 
 #### Generate necessary symlinks and copying files ####
 echo 'We are now going to generate some symlinks.'
 read -p 'Press any key to continue or ctl-c to exit'
+sudo ln -s /usr/local/BerkeleyDB.5.0/lib/libdb-5.0.so /usr/lib64/libdb-5.0.so
+sudo ln -s /usr/lib64/mysql/libmysqlclient.so /usr/lib64/libmysqlclient.so
+sudo ln -s /usr/lib64/mysql/libmysqlclient.so.16 /usr/lib64/libmysqlclient.so.16
+sudo ldconfig
 ln -s ~/git/PublicEngine/MMOEngine ~/git/Core3/MMOEngine
 sudo ln -s ~/git/PublicEngine/MMOEngine/bin/idlc /usr/local/bin/idlc
 chmod 755 ~/git/PublicEngine/MMOEngine/bin/idlc