+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * 7/10/15 - Tested firstrun.sh & build.sh, made some slight changes to both
 * 7/14/15 - firstrun.sh is good to go. build.sh is failing with linking error for mysqlclient 
+* 7/26/15 - firstrun.sh, build.sh and webmin.sh all tested and working